+++ packages/common/src/cognis/__init__.py
@@ -6,4 +6,4 @@ from pkgutil import extend_path
 
 __path__ = extend_path(__path__, __name__)
 
-__version__ = "0.14.1"
+__version__ = "0.14.2"